requirements.txt
python-dotenv==0.13.0
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
asyncpg